Removed SimpleMethodInfo and replaced by ClassInfoAware
